+++ src/com/reco1l/andengine/buffered/BufferedEntity.kt
@@ -1,6 +1,7 @@
 package com.reco1l.andengine.buffered
 
 import com.reco1l.andengine.*
+import com.reco1l.andengine.info.*
 import org.anddev.andengine.engine.camera.Camera
 import org.anddev.andengine.entity.shape.Shape.*
 import org.anddev.andengine.opengl.util.*
