+++ src/com/reco1l/andengine/info/BlendInfo.kt
@@ -1,4 +1,4 @@
-package com.reco1l.andengine
+package com.reco1l.andengine.info
 
 import javax.microedition.khronos.opengles.*
