+++ src/com/reco1l/andengine/info/ClearInfo.kt
@@ -1,4 +1,4 @@
-package com.reco1l.andengine
+package com.reco1l.andengine.info
 
 data class ClearInfo(
